+++ src/frpchttpclient.h
@@ -20,7 +20,7 @@
  * Radlicka 2, Praha 5, 15000, Czech Republic
  * http://www.seznam.cz, mailto:fastrpc@firma.seznam.cz
  *
- * FILE          $Id: frpchttpclient.h,v 1.2 2007-04-02 15:28:20 vasek Exp $
+ * FILE          $Id: frpchttpclient.h,v 1.3 2007-05-21 15:10:12 mirecta Exp $
  *
  * DESCRIPTION   
  *
@@ -57,26 +57,26 @@ class FRPC_DLLEXPORT DataSink_t
 {
 public:
     inline DataSink_t(UnMarshaller_t& um,
-                      long type = UnMarshaller_t::TYPE_METHOD_RESPONSE )
+                      unsigned int type = UnMarshaller_t::TYPE_METHOD_RESPONSE )
                       :um(um),dataWritten(0), type(type)
     {}
 
     inline ~DataSink_t()
     {}
-    inline void write(const char *data, long size)
+    inline void write(const char *data, unsigned int size)
     {
         um.unMarshall(data, size, type);
         dataWritten += size;
     }
-    inline unsigned long written()
+    inline unsigned int written()
     {
         return dataWritten;
     }
 
 private:
     UnMarshaller_t &um;
-    unsigned long dataWritten;
-    long type;
+    unsigned int dataWritten;
+    unsigned int type;
 };
 
 /**
@@ -86,10 +86,10 @@ class FRPC_DLLEXPORT HTTPClient_t : public Writer_t
 {
 public:
     
-    HTTPClient_t(HTTPIO_t &httpIO, URL_t &url, long connectTimeout,
+    HTTPClient_t(HTTPIO_t &httpIO, URL_t &url, unsigned int connectTimeout,
                  bool keepAlive);
 
-    HTTPClient_t(HTTPIO_t &httpIO, URL_t &url, long connectTimeout,
+    HTTPClient_t(HTTPIO_t &httpIO, URL_t &url, unsigned int connectTimeout,
                  bool keepAlive, bool useHTTP10);
 
 
@@ -102,7 +102,7 @@ public:
     * @param contentType maybe  XML_RPC or BINARY_RPC
     */
 
-    inline void prepare(unsigned long contentType)
+    inline void prepare(unsigned int contentType)
     {
         switch(contentType)
         {
@@ -124,7 +124,7 @@ public:
     * @return maybe XML_RPC , BINARY_RPC or both
     */
 
-    inline unsigned long getSupportedProtocols()
+    inline unsigned int getSupportedProtocols()
     {
         return supportedProtocols;
     }
@@ -138,7 +138,7 @@ public:
     * @param data pointer to data 
     * @param size size of data
     */
-    virtual void write(const char* data, long size);
+    virtual void write(const char* data, unsigned int size);
     /**
     *@brief read response from socket and unmarshaling
     *@param builder is a builder required for unmarshaller to build data tree
@@ -152,15 +152,15 @@ private:
 
     HTTPIO_t &httpIO;
     URL_t    &url;
-    long connectTimeout;
+    unsigned int connectTimeout;
     bool keepAlive;
 
 
     bool headersSent;
     bool useChunks;
-    unsigned long supportedProtocols;
-    unsigned long useProtocol;
-    unsigned long contentLenght;
+    unsigned int supportedProtocols;
+    unsigned int useProtocol;
+    unsigned int contentLenght;
     bool connectionMustClose;
 
     std::list<std::string> queryStorage;
