+++ src/frpcmethodregistry.cc
@@ -20,7 +20,7 @@
  * Radlicka 2, Praha 5, 15000, Czech Republic
  * http://www.seznam.cz, mailto:fastrpc@firma.seznam.cz
  *
- * FILE          $Id: frpcmethodregistry.cc,v 1.8 2007-05-18 15:29:45 mirecta Exp $
+ * FILE          $Id: frpcmethodregistry.cc,v 1.9 2007-05-21 15:10:12 mirecta Exp $
  *
  * DESCRIPTION   
  *
@@ -163,7 +163,7 @@ namespace
 const long BUFFER_SIZE = 1<<16;
 }
 */
-long MethodRegistry_t::processCall(const std::string &clientIP, const std::string &methodName,
+int MethodRegistry_t::processCall(const std::string &clientIP, const std::string &methodName,
                                    Array_t &params,
                                    Writer_t &writer, unsigned int typeOut,
                                    const ProtocolVersion_t &protocolVersion)
@@ -316,7 +316,7 @@ Value_t& MethodRegistry_t::processCall(const std::string &clientIP, Reader_t &re
     return *result;
 }
 
-long MethodRegistry_t::processCall(const std::string &clientIP, Reader_t &reader,
+int MethodRegistry_t::processCall(const std::string &clientIP, Reader_t &reader,
                                    unsigned int typeIn, Writer_t &writer, 
                                    unsigned int typeOut)
 {
@@ -327,7 +327,7 @@ long MethodRegistry_t::processCall(const std::string &clientIP, Reader_t &reader
     
 
     char buffer[BUFFER_SIZE];
-    long readed;
+    unsigned int readed;
 
 
     try
@@ -372,7 +372,7 @@ long MethodRegistry_t::processCall(const std::string &clientIP, Reader_t &reader
     return 0;
 }
 
-long MethodRegistry_t::headCall()
+int MethodRegistry_t::headCall()
 {
 
     //if head method registered call it
